Production ready: Read description
+ user-account/edituser
+ prepare-for-change/editprepareforchange
+ know-yourself/editknowyourself
+ explore-opportunities/editexploreopportunities
+ user-account/loginuser: This is an endpoint that I've designed and implemented all by myself. It wasn't originally in the design, but I've updated both spreadsheets to document its design. Approach me for assistance.

This is the build I'm confident in bringing into production. It includes everything that was planned for, in addition to an additional endpoint associated with logging in. A lot of effort and consideration went into the making, but I'd like to thank Zhong Wei for helping me on the research while I couldn't on the first 3 days, and I'm glad I was able to get it all done by myself.

As much as I try to compensate for edge cases and security concerns, there are some that will either slip through my mind, or will be too hard for me to catch. These I wish to discuss in the next meeting with Mr. Joe. Until then, here are all the endpoints, done to my best effort, and fully usable. As always, I'm available for any tech support. Best of luck, frontend.